hotfix : Fix delete course

See merge request miw-se-cursisten/cohort-apr24/quizmaster/team-1/quizmaster!91

+++ src/main/java/controller/ManageCoursesController.java
@@ -135,6 +135,10 @@ public class ManageCoursesController {
         }
 
         if (confirmMessage("Course Verwijderen",message )) {
+            if(selectedCourse.getStudentCount()>0) {
+                showAlert(Alert.AlertType.ERROR, "Fout", "Cursus heeft studenten, kan niet worden verwijderd");
+                return;
+            }
             courseDao.deleteOneById(selectedCourse.getCourseId());
             courseTable.getItems().remove(selectedCourse);
         }
